+++ src/pages/account/signup.tsx
@@ -1,28 +1,14 @@
 import AccountLayout from '@/components/layouts/account'
+import SignupForm from '@/components/pages/account/SignupForm'
 import LockOutlinedIcon from '@mui/icons-material/LockOutlined'
 import {
   Avatar,
   Box,
-  Button,
-  Checkbox,
-  FormControlLabel,
-  Grid,
-  Link,
-  TextField,
-  Typography,
+  Typography
 } from '@mui/material'
 import { ReactElement } from 'react'
 
 export default function SignUp() {
-  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault()
-    const data = new FormData(event.currentTarget)
-    console.log({
-      email: data.get('email'),
-      password: data.get('password'),
-    })
-  }
-
   return (
     <Box
       sx={{
@@ -38,76 +24,7 @@ export default function SignUp() {
       <Typography component="h1" variant="h5">
         アカウントを作成
       </Typography>
-      <Box component="form" noValidate onSubmit={handleSubmit} sx={{ mt: 3 }}>
-        <Grid container spacing={2}>
-          <Grid item xs={12} sm={6}>
-            <TextField
-              autoComplete="given-name"
-              name="firstName"
-              required
-              fullWidth
-              id="firstName"
-              label="姓"
-              autoFocus
-            />
-          </Grid>
-          <Grid item xs={12} sm={6}>
-            <TextField
-              required
-              fullWidth
-              id="lastName"
-              label="名"
-              name="lastName"
-              autoComplete="family-name"
-            />
-          </Grid>
-          <Grid item xs={12}>
-            <TextField
-              required
-              fullWidth
-              id="email"
-              label="メールアドレス"
-              name="email"
-              autoComplete="email"
-            />
-          </Grid>
-          <Grid item xs={12}>
-            <TextField
-              required
-              fullWidth
-              name="password"
-              label="パスワード"
-              type="password"
-              id="password"
-              autoComplete="new-password"
-            />
-          </Grid>
-          <Grid item xs={12}>
-            <FormControlLabel
-              control={<Checkbox value="allowExtraEmails" color="primary" />}
-              label="利用規約とプライバシーポリシーに同意する"
-            />
-          </Grid>
-        </Grid>
-        <Button
-          type="submit"
-          fullWidth
-          variant="contained"
-          sx={{ mt: 3, mb: 2 }}
-        >
-          サインアップ
-        </Button>
-
-        <Grid container justifyContent="flex-end">
-          <Grid item>
-            すでにアカウントをお持ちですか？
-            <Link href="#" variant="body2">
-              ログイン
-            </Link>
-            する
-          </Grid>
-        </Grid>
-      </Box>
+      <SignupForm />
     </Box>
   )
 }
